# frontend/openapi.json
{
  "components": {
    "schemas": {
      "ApprovalRequest": {
        "description": "The HITL spine \u2014 every human gate in any workflow passes through one of these.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "graph_thread_id": {
            "title": "Graph Thread Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus",
            "default": "pending"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "workflow": {
            "title": "Workflow",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow",
          "graph_thread_id",
          "kind",
          "payload",
          "created_at"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "ApprovalStatus": {
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
...
      "CallRecord": {
        "description": "A voice call, raw and extracted layers kept separate.",
        "properties": {
          "contact_id": {
            "default": "",
            "title": "Contact Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "extracted": {
            "additionalProperties": true,
            "title": "Extracted",
            "type": "object"
          },
          "listing_key": {
            "default": "",
            "title": "Listing Key",
            "type": "string"
          },
          "outcome": {
            "default": "",
            "title": "Outcome",
            "type": "string"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "transcript": {
            "default": "",
            "title": "Transcript",
            "type": "string"
          },
          "vapi_call_id": {
            "title": "Vapi Call Id",
            "type": "string"
          }
        },
        "required": [
          "vapi_call_id"
        ],
        "title": "CallRecord",
        "type": "object"
      },
      "Contact": {
        "description": "CRM contact DTO. The CRM behind `CRMPort` is the source of truth \u2014\nthis is a read-through shape, never persisted on our side.\n\n`crm_id` is the contact's id in whichever CRM this deploy is wired to\n(a FollowUpBoss person id, a Sierra Interactive lead id, \u2026); it is only\nmeaningful to the adapter that produced it.",
        "properties": {
          "crm_id": {
            "title": "Crm Id",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "default": "Lead",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "crm_id",
          "name"
        ],
        "title": "Contact",
        "type": "object"
      },
...
      "DecideRequest": {
        "description": "Wire shape for a decision. `decided_by` is intentionally absent \u2014 the API\nstamps it from the authenticated principal, never the client.\n\n`edited_payload` stays a dict here because its shape depends on the approval\n*kind*, known only after the row is fetched; the decide route validates it\nagainst the kind's boundary model (EditedMessagePayload for the\noutbound-message gate) before anything reaches an engine.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "edited_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Payload"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecideRequest",
        "type": "object"
      },
      "DecisionResponse": {
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/ApprovalRequest"
          },
          "workflow": {
            "$ref": "#/components/schemas/WorkflowRunResult"
          }
        },
        "required": [
          "approval",
          "workflow"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "FeedbackSource": {
        "enum": [
          "call",
          "form"
        ],
        "title": "FeedbackSource",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutboundCallRequest": {
        "properties": {
          "contact_id": {
            "title": "Contact Id",
            "type": "string"
          },
          "listing_key": {
            "title": "Listing Key",
            "type": "string"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          }
        },
        "required": [
          "listing_key",
          "contact_id"
        ],
        "title": "OutboundCallRequest",
        "type": "object"
      },
      "OutboundCallResult": {
        "properties": {
          "call_id": {
            "title": "Call Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "call_id"
        ],
        "title": "OutboundCallResult",
        "type": "object"
      },
...
      "Sentiment": {
        "enum": [
          "positive",
          "neutral",
          "negative"
        ],
        "title": "Sentiment",
        "type": "string"
      },
...
      "ShowingFeedback": {
        "properties": {
          "call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Id"
          },
          "contact_id": {
            "title": "Contact Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing_key": {
            "title": "Listing Key",
            "type": "string"
          },
          "sentiment": {
            "$ref": "#/components/schemas/Sentiment",
            "default": "neutral"
          },
          "source": {
            "$ref": "#/components/schemas/FeedbackSource",
            "default": "call"
          },
          "structured_answers": {
            "additionalProperties": true,
            "title": "Structured Answers",
            "type": "object"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_key",
          "contact_id"
        ],
        "title": "ShowingFeedback",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowRunResult": {
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "status"
        ],
        "title": "WorkflowRunResult",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "brokerops api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/approvals": {
      "get": {
        "description": "List approvals, newest first \u2014 role-filtered on every path (BOP-040).\n\nDefault: the pending inbox. `transaction_id` returns one deal's full approval\nhistory for the transaction hub (BOP-027) \u2014 pending AND decided, so `status` is\nnot applied in that mode. Both are caller-role filtered: a viewer receives each\ngate's kind/status but the draft payload (recipient/subject/body of an\noutbound-message gate) is redacted to null, so a viewer-open inbox never delivers\ndraft content over the wire; an operator/admin sees it intact. The transaction id\nlives in the open payload, so the transaction filter reads it before scrubbing.",
        "operationId": "list_approvals_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "transaction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Transaction Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRequest"
                  },
                  "title": "Response List Approvals Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Approvals",
        "tags": [
          "approvals"
        ]
      }
    },
    "/approvals/{approval_id}": {
      "get": {
        "operationId": "get_approval_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval",
        "tags": [
          "approvals"
        ]
      }
    },
    "/approvals/{approval_id}/decide": {
      "post": {
        "operationId": "decide_approval_approvals__approval_id__decide_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Approval",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/calls/outbound": {
      "post": {
        "operationId": "start_outbound_call_calls_outbound_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutboundCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutboundCallResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Outbound Call",
        "tags": [
          "calls"
        ]
      }
    },
    "/calls/{call_id}": {
      "get": {
        "operationId": "get_call_record_calls__call_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "call_id",
            "required": true,
            "schema": {
              "title": "Call Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Call Record",
        "tags": [
          "calls"
        ]
      }
    },
    "/contacts": {
      "get": {
        "operationId": "search_contacts_contacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  },
                  "title": "Response Search Contacts Contacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Contacts",
        "tags": [
          "contacts"
        ]
      }
    },
    "/contacts/{contact_id}": {
      "get": {
        "operationId": "get_contact_contacts__contact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contact_id",
            "required": true,
            "schema": {
              "title": "Contact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contact",
        "tags": [
          "contacts"
        ]
      }
    },
...
    "/feedback": {
      "get": {
        "operationId": "list_feedback_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "listing_key",
            "required": true,
            "schema": {
              "title": "Listing Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShowingFeedback"
                  },
                  "title": "Response List Feedback Feedback Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feedback",
        "tags": [
          "calls"
        ]
      }
    },
...
  }
}